storage/requirements.txt
connexion
swagger-ui-bundle
SQLAlchemy
mysql-connector-python
pymysql
pykafka
flask
cryptography
flask_cors
